API Call in Upload Interactor
The information displayed about the user's uploaded image is now relevant, rather
than boilerplate text. The API call is tested for success, and on failure notifies
the user through a popup dialog that their call did not go through.

+++ src/main/java/use_case/upload/UploadResultOutputData.java
@@ -5,9 +5,10 @@ public class UploadResultOutputData {
     private final String name;
     private final String scientificName;
     private final String family;
-    private final int certainty;
+    private final double certainty;
 
-    public UploadResultOutputData(String imagePath, String name, String scientificName, String family, int certainty) {
+    public UploadResultOutputData(String imagePath, String name, String scientificName,
+                                  String family, double certainty) {
         this.imagePath = imagePath;
         this.name = name;
         this.scientificName = scientificName;
@@ -31,7 +32,7 @@ public class UploadResultOutputData {
         return family;
     }
 
-    public int getCertainty() {
+    public double getCertainty() {
         return certainty;
     }
 }
